+++ src/test/java/io/xpring/xrpl/DefaultXRPClientTest.java
@@ -101,7 +101,7 @@ public class DefaultXRPClientTest {
 
 
     @Test
-    public void transactionStatusWithUnvalidatedTransactionAndFailureCode() throws IOException, XpringException {
+    public void paymentStatusWithUnvalidatedTransactionAndFailureCode() throws IOException, XpringException {
         // Iterate over different types of transaction status codes which represent failures.
         for (String transactionFailureCode : TRANSACTION_FAILURE_STATUS_CODES) {
             // GIVEN a XRPClient which will return an unvalidated transaction with a failed code.
@@ -112,16 +112,16 @@ public class DefaultXRPClientTest {
                     GRPCResult.ok(makeSubmitTransactionResponse(TRANSACTION_HASH))
             );
 
-            // WHEN the transaction status is retrieved.
-            io.xpring.xrpl.TransactionStatus transactionStatus = client.getTransactionStatus(TRANSACTION_HASH);
+            // WHEN the payment status is retrieved.
+            io.xpring.xrpl.TransactionStatus paymentStatus = client.getPaymentStatus(TRANSACTION_HASH);
 
             // THEN the status is PENDING.
-            assertThat(transactionStatus).isEqualTo(io.xpring.xrpl.TransactionStatus.PENDING);
+            assertThat(paymentStatus).isEqualTo(io.xpring.xrpl.TransactionStatus.PENDING);
         }
     }
 
     @Test
-    public void transactionStatusWithUnvalidatedTransactionAndSuccessCode() throws IOException, XpringException {
+    public void paymentStatusWithUnvalidatedTransactionAndSuccessCode() throws IOException, XpringException {
         // GIVEN a XRPClient which will return an unvalidated transaction with a success code.
         DefaultXRPClient client = getClient(
                 GRPCResult.ok(makeGetAccountInfoResponse(DROPS_OF_XRP_IN_ACCOUNT)),
@@ -130,15 +130,15 @@ public class DefaultXRPClientTest {
                 GRPCResult.ok(makeSubmitTransactionResponse(TRANSACTION_HASH))
         );
 
-        // WHEN the transaction status is retrieved.
-        io.xpring.xrpl.TransactionStatus transactionStatus = client.getTransactionStatus(TRANSACTION_HASH);
+        // WHEN the payment status is retrieved.
+        io.xpring.xrpl.TransactionStatus paymentStatus = client.getPaymentStatus(TRANSACTION_HASH);
 
         // THEN the status is PENDING.
-        assertThat(transactionStatus).isEqualTo(io.xpring.xrpl.TransactionStatus.PENDING);
+        assertThat(paymentStatus).isEqualTo(io.xpring.xrpl.TransactionStatus.PENDING);
     }
 
     @Test
-    public void transactionStatusWithValidatedTransactionAndFailureCode() throws IOException, XpringException {
+    public void paymentStatusWithValidatedTransactionAndFailureCode() throws IOException, XpringException {
         // Iterate over different types of transaction status codes which represent failures.
         for (String transactionFailureCode : TRANSACTION_FAILURE_STATUS_CODES) {
             // GIVEN a XRPClient which will return an validated transaction with a failed code.
@@ -149,16 +149,16 @@ public class DefaultXRPClientTest {
                     GRPCResult.ok(makeSubmitTransactionResponse(TRANSACTION_HASH))
             );
 
-            // WHEN the transaction status is retrieved.
-            io.xpring.xrpl.TransactionStatus transactionStatus = client.getTransactionStatus(TRANSACTION_HASH);
+            // WHEN the payment status is retrieved.
+            io.xpring.xrpl.TransactionStatus paymentStatus = client.getPaymentStatus(TRANSACTION_HASH);
 
             // THEN the status is FAILED.
-            assertThat(transactionStatus).isEqualTo(io.xpring.xrpl.TransactionStatus.FAILED);
+            assertThat(paymentStatus).isEqualTo(io.xpring.xrpl.TransactionStatus.FAILED);
         }
     }
 
     @Test
-    public void transactionStatusWithValidatedTransactionAndSuccessCode() throws IOException, XpringException {
+    public void paymentStatusWithValidatedTransactionAndSuccessCode() throws IOException, XpringException {
         // GIVEN a XRPClient which will return an validated transaction with a success code.
         DefaultXRPClient client = getClient(
                 GRPCResult.ok(makeGetAccountInfoResponse(DROPS_OF_XRP_IN_ACCOUNT)),
@@ -167,15 +167,15 @@ public class DefaultXRPClientTest {
                 GRPCResult.ok(makeSubmitTransactionResponse(TRANSACTION_HASH))
         );
 
-        // WHEN the transaction status is retrieved.
-        io.xpring.xrpl.TransactionStatus transactionStatus = client.getTransactionStatus(TRANSACTION_HASH);
+        // WHEN the payment status is retrieved.
+        io.xpring.xrpl.TransactionStatus paymentStatus = client.getPaymentStatus(TRANSACTION_HASH);
 
         // THEN the status is SUCCEEDED.
-        assertThat(transactionStatus).isEqualTo(io.xpring.xrpl.TransactionStatus.SUCCEEDED);
+        assertThat(paymentStatus).isEqualTo(io.xpring.xrpl.TransactionStatus.SUCCEEDED);
     }
 
     @Test
-    public void transactionStatusWithNodeError() throws IOException, XpringException {
+    public void paymentStatusWithNodeError() throws IOException, XpringException {
         // GIVEN a XRPClient which will error when a transaction status is requested..
         DefaultXRPClient client = getClient(
                 GRPCResult.ok(makeGetAccountInfoResponse(DROPS_OF_XRP_IN_ACCOUNT)),
@@ -184,9 +184,9 @@ public class DefaultXRPClientTest {
                 GRPCResult.ok(makeSubmitTransactionResponse(TRANSACTION_HASH))
         );
 
-        // WHEN the transaction status is retrieved THEN an error is thrown..
+        // WHEN the payment status is retrieved THEN an error is thrown..
         expectedException.expect(Exception.class);
-        client.getTransactionStatus(TRANSACTION_HASH);
+        client.getPaymentStatus(TRANSACTION_HASH);
     }
 
     @Test